**Summarizing the Diff**
I'm currently focusing on crafting a concise commit message for a new JPEG image addition. The diff reveals a file named `A9B1A5AE-6031-4E5A-ACF5-0 0C58432F38B_4_5005_c 2.jpeg` is added to the project. My goal is to capture this change succinctly.

**Refining the Message**

I've honed in on the best phrasing. While the original file name is lengthy, it's the addition to the `pic` image set that matters most. Considering the goal for brevity, my preferred commit message is now finalized: "Add `pic` image asset". It precisely captures the core of the diff.

**Formulating the Commit Message**

I'm solidifying the language further. Based on the diff analysis, the core action is adding a JPEG to the `pic` image set. The filename is less important than the context. The examples provided helped refine the message. I'm set on keeping the message succinct and descriptive of the action: adding an image asset to the `pic` image set. The message I will use is "Add `pic` image asset".

Add `pic` image asset